Added article author name
The new code segment addresses the author of the article. As you can
see, this queries the 'user' table and returns the actual user name of
the author. This should probably be changed so it uses com_content
instead. This way it could also include the author alias. I just didn't
have time to delve into that yet.